fix: stop saturating the summary's latency distribution at int32
729002d widened the recorded latency to long, and an end-to-end run then
showed the fix was only half of one. A CUBRID TPC-H Q20 at SF=0.1 wrote
2594811975 us -- 43.2 minutes, correct -- into the raw CSV, while the
summary.json beside it reported every single percentile as exactly
2147483647:

  "Median Latency (microseconds)":  2147483647,
  "Average Latency (microseconds)": 2147483647,
  "Maximum Latency (microseconds)": 2147483647,

That is Integer.MAX_VALUE, and it is saturation rather than the wrap the
recording path had. DistributionStatistics keeps percentiles in a long but
hands them out through double getters, and toMap() cast those doubles to
int. A double-to-int cast clamps to Integer.MAX_VALUE where a long-to-int
cast would have wrapped negative, so the summary looked like an ordinary
35.8-minute answer with nothing to give it away -- and unlike the negative
values in the raw CSV, nothing downstream could detect it.

Widen toMap() to Map<String, Long>. Its one caller puts the result into a
Map<String, Object>, so no other signature moves.

The test pins all three behaviours apart, since which cast is used decides
whether a too-large latency comes back wrong-and-visible or wrong-and-silent.

+++ src/main/java/com/oltpbenchmark/DistributionStatistics.java
@@ -169,17 +169,23 @@ public class DistributionStatistics {
         + "]";
   }
 
-  public Map<String, Integer> toMap() {
-    Map<String, Integer> distMap = new LinkedHashMap<>();
-    distMap.put("Minimum Latency (microseconds)", (int) getMinimum());
-    distMap.put("25th Percentile Latency (microseconds)", (int) get25thPercentile());
-    distMap.put("Median Latency (microseconds)", (int) getMedian());
-    distMap.put("Average Latency (microseconds)", (int) getAverage());
-    distMap.put("75th Percentile Latency (microseconds)", (int) get75thPercentile());
-    distMap.put("90th Percentile Latency (microseconds)", (int) get90thPercentile());
-    distMap.put("95th Percentile Latency (microseconds)", (int) get95thPercentile());
-    distMap.put("99th Percentile Latency (microseconds)", (int) get99thPercentile());
-    distMap.put("Maximum Latency (microseconds)", (int) getMaximum());
+  /**
+   * Latencies are microseconds in a long, because a single analytical query can exceed the 35.8
+   * minutes that fit in an int. The casts below used to be to int, and a double-to-int cast
+   * saturates rather than wraps, so every percentile of a slow TPC-H stream reported exactly
+   * Integer.MAX_VALUE while the raw CSV beside it held the true value.
+   */
+  public Map<String, Long> toMap() {
+    Map<String, Long> distMap = new LinkedHashMap<>();
+    distMap.put("Minimum Latency (microseconds)", (long) getMinimum());
+    distMap.put("25th Percentile Latency (microseconds)", (long) get25thPercentile());
+    distMap.put("Median Latency (microseconds)", (long) getMedian());
+    distMap.put("Average Latency (microseconds)", (long) getAverage());
+    distMap.put("75th Percentile Latency (microseconds)", (long) get75thPercentile());
+    distMap.put("90th Percentile Latency (microseconds)", (long) get90thPercentile());
+    distMap.put("95th Percentile Latency (microseconds)", (long) get95thPercentile());
+    distMap.put("99th Percentile Latency (microseconds)", (long) get99thPercentile());
+    distMap.put("Maximum Latency (microseconds)", (long) getMaximum());
     return distMap;
   }
 }
